+++ src/modules/mail-manager/mail-manager.module.ts
@@ -3,6 +3,8 @@ import { MailerModule } from '@nestjs-modules/mailer';
 import { MailManagerService } from './mail-manager.service';
 import { MailerAsyncOptions } from '@nestjs-modules/mailer/dist/interfaces/mailer-async-options.interface';
 import { LoggerModule } from '../logger/logger.module';
+import { HandlebarsAdapter } from '@nestjs-modules/mailer/dist/adapters/handlebars.adapter';
+import { join } from 'path';
 
 @Module({
     imports: [
@@ -14,6 +16,13 @@ import { LoggerModule } from '../logger/logger.module';
                     auth: {
                         user: process.env.MAIL_USER,
                         pass: process.env.MAIL_PASS
+                    },
+                },
+                template: {
+                    dir: join(process.cwd(), 'templates'),
+                    adapter: new HandlebarsAdapter(),
+                    options: {
+                        strict: true,
                     },
                 }
             })