+++ src/modules/mail-manager/mail-manager.service.ts
@@ -10,25 +10,26 @@ export class MailManagerService {
         private readonly loggerService: LoggerService
     ) {}
 
-    async sendActivationMail(to: string, activationCode: string): Promise<void> {
-        await this.sendMail(to, 'Verification', [
-            'Thanks for your registration!',
-            `To activate your account, use this code: ${activationCode}.`
-        ]);
+    async sendActivationMail(to: string, username: string, activationCode: string): Promise<void> {
+        await this.sendMail(to, 'Witamy na naszej platformie!', username, activationCode);
     }
 
-    private async sendMail(to: string, subject: string, contentArr: string[]): Promise<void> {
+    private async sendMail(to: string, subject: string, username: string, activationCode: string): Promise<void> {
         const context = 'MailManagerService/sendMail';
 
         try {
-            const result = await this.mailerService.sendMail({
+            const { accepted, messageTime } = await this.mailerService.sendMail({
                 to,
                 from: process.env.MAIL_ADDRESS,
                 subject,
-                text: contentArr.join(';')
+                context: {
+                    username,
+                    link: `${process.env.FRONTEND_WEB}/users/${activationCode}/activate`
+                },
+                template: 'activation'
             });
 
-            this.loggerService.info(context, result);
+            this.loggerService.info(context, `Send activation mail to "${accepted[0]}". It took ${(messageTime / 1000).toFixed(1)}s`);
         } catch (err) {
             this.loggerService.error(context, err);
         }