+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.0] - 2024-10-21
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): `Handlebars` library for templates.
+- [Marek Kurańda](https://github.com/mjkuranda): Template for activation mail.
+- [Marek Kurańda](https://github.com/mjkuranda): `FRONTEND_WEB`, `FRONTEND_MOBILE` and `FRONTEND_OTHERS` variables.
+
+### Changed
+- [Marek Kurańda](https://github.com/mjkuranda): Mail verification content.
+
 ## [2.4.0] - 2024-10-15
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Throwing an error when dish is marked as `soft-deleted`.